+++ Лаб 2.5/Лаб 2.5/Moped.cs
@@ -1,7 +1,7 @@
 ﻿class Moped : TwoWheeledTransport
 {   
     public string Basket {  get; set; }
-    public Moped(string type, int passengerCapacity, double cargoCapacity, int price, int speed)
+    public Moped(string type, int passengerCapacity, int cargoCapacity, int price, int speed)
         : base(type, passengerCapacity, cargoCapacity, price, speed)
     {
     }
